Move defaultProps to separate props file
