+++ src/players/YouTube.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import loadScript from 'load-script'
 
-import propTypes from '../propTypes'
+import { propTypes, defaultProps } from '../props'
 import Base from './Base'
 
 const SDK_URL = '//www.youtube.com/iframe_api'
@@ -16,9 +16,7 @@ const DEFAULT_PLAYER_VARS = {
 
 export default class YouTube extends Base {
   static propTypes = propTypes
-  static defaultProps = {
-    youtubeConfig: {}
-  }
+  static defaultProps = defaultProps
   static canPlay (url) {
     return MATCH_URL.test(url)
   }
