+++ src/props.js
@@ -1,6 +1,6 @@
 import { PropTypes } from 'react'
 
-export default {
+export const propTypes = {
   url: PropTypes.string,
   playing: PropTypes.bool,
   volume: PropTypes.number,
@@ -20,4 +20,26 @@ export default {
   onBuffer: PropTypes.func,
   onEnded: PropTypes.func,
   onError: PropTypes.func
+}
+
+export const defaultProps = {
+  playing: false,
+  width: 640,
+  height: 360,
+  volume: 0.8,
+  soundcloudConfig: {
+    clientId: 'e8b6f84fbcad14c301ca1355cae1dea2'
+  },
+  youtubeConfig: {
+    playerVars: {}
+  },
+  vimeoConfig: {
+    iframeParams: {}
+  },
+  onPlay: function () {},
+  onPause: function () {},
+  onBuffer: function () {},
+  onEnded: function () {},
+  onError: function () {},
+  onProgress: function () {}
 }
